requirements.txt
Django==1.8.4
django-extensions==1.6.1
djangorestframework==3.3.2
redis==2.10.5
six==1.10.0
wheel==0.24.0
